+++ extension/chrome/elements/compose.ts
@@ -29,7 +29,7 @@ import { ComposeSizeModule } from './compose-modules/compose-size-module.js';
 import { ComposeStorageModule } from './compose-modules/compose-storage-module.js';
 import { Catch } from '../../js/common/platform/catch.js';
 import { Rules } from '../../js/common/rules.js';
-import { Keyserver } from '../../js/common/api/keyserver.js';
+import { PubLookup } from '../../js/common/api/pub-lookup.js';
 import { Scopes, AcctStore } from '../../js/common/platform/store/acct-store.js';
 
 export class ComposeView extends View {
@@ -54,7 +54,7 @@ export class ComposeView extends View {
   public replyParams: ReplyParams | undefined;
   public emailProvider: EmailProviderInterface;
   public rules!: Rules;
-  public keyserver!: Keyserver;
+  public pubLookup!: PubLookup;
 
   public quoteModule!: ComposeQuoteModule;
   public sendBtnModule!: ComposeSendBtnModule;
@@ -142,7 +142,7 @@ export class ComposeView extends View {
   public render = async () => {
     const storage = await AcctStore.get(this.acctEmail, ['sendAs', 'hide_message_password', 'drafts_reply']);
     this.rules = await Rules.newInstance(this.acctEmail);
-    this.keyserver = new Keyserver(this.rules);
+    this.pubLookup = new PubLookup(this.rules);
     this.tabId = await BrowserMsg.requiredTabId();
     this.factory = new XssSafeFactory(this.acctEmail, this.tabId);
     this.scopes = await AcctStore.getScopes(this.acctEmail);
