+++ extension/chrome/settings/modules/my_key_update.ts
@@ -14,7 +14,7 @@ import { Url, Str } from '../../../js/common/core/common.js';
 import { View } from '../../../js/common/view.js';
 import { opgp } from '../../../js/common/core/pgp.js';
 import { Rules } from '../../../js/common/rules.js';
-import { Keyserver } from '../../../js/common/api/keyserver.js';
+import { PubLookup } from '../../../js/common/api/pub-lookup.js';
 import { KeyStore } from '../../../js/common/platform/store/key-store.js';
 import { PassphraseStore } from '../../../js/common/platform/store/passphrase-store.js';
 
@@ -26,7 +26,7 @@ View.run(class MyKeyUpdateView extends View {
   private readonly prvHeaders = PgpArmor.headers('privateKey');
   private primaryKi: KeyInfo | undefined;
   private rules!: Rules;
-  private keyserver!: Keyserver;
+  private pubLookup!: PubLookup;
 
   constructor() {
     super();
@@ -38,7 +38,7 @@ View.run(class MyKeyUpdateView extends View {
 
   public render = async () => {
     this.rules = await Rules.newInstance(this.acctEmail);
-    this.keyserver = new Keyserver(this.rules);
+    this.pubLookup = new PubLookup(this.rules);
     [this.primaryKi] = await KeyStore.get(this.acctEmail, [this.longid]);
     Assert.abortAndRenderErrorIfKeyinfoEmpty(this.primaryKi);
     $('.action_show_public_key').attr('href', this.showKeyUrl);
@@ -59,7 +59,7 @@ View.run(class MyKeyUpdateView extends View {
     await PassphraseStore.set('session', this.acctEmail, this.primaryKi!.longid, typeof storedPassphrase !== 'undefined' ? undefined : updatedPrvPassphrase);
     if (this.rules.canSubmitPubToAttester() && await Ui.modal.confirm('Public and private key updated locally.\n\nUpdate public records with new Public Key?')) {
       try {
-        await Ui.modal.info(await this.keyserver.attester.updatePubkey(this.primaryKi!.longid, updatedPrv.toPublic().armor()));
+        await Ui.modal.info(await this.pubLookup.attester.updatePubkey(this.primaryKi!.longid, updatedPrv.toPublic().armor()));
       } catch (e) {
         ApiErr.reportIfSignificant(e);
         await Ui.modal.error(`Error updating public records:\n\n${ApiErr.eli5(e)}\n\n(but local update was successful)`);
