+++ extension/js/common/api/sks.ts
@@ -5,7 +5,7 @@
 import { Api } from './api.js';
 import { ApiErr } from './error/api-error.js';
 import { PgpArmor } from '../core/pgp-armor.js';
-import { PubkeySearchResult } from './keyserver.js';
+import { PubkeySearchResult } from './pub-lookup.js';
 
 export class Sks extends Api {
